Server/userServices/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
